refactor: WindowsContext path
