+++ src/view/features/popup/components/SearchResult.tsx
@@ -12,8 +12,8 @@ import {
   focusTab,
   getRecentActiveTabs,
 } from "../../../../repository/TabsRepository";
+import { WindowsContext } from "../../../contexts/Windows";
 import TabItem from "../components/TabItem";
-import { WindowsContext } from "../contexts/Windows";
 
 type SearchResultProps = {
   searchText: string;
