+++ src/view/features/popup/components/TabList.tsx
@@ -8,7 +8,7 @@ import { useContext } from "react";
 import { Tab } from "../../../../model/Tab";
 import { isPinned, isTabGroup } from "../../../../model/TabContainer";
 import { WindowChild } from "../../../../model/Window";
-import { WindowsContext } from "../contexts/Windows";
+import { WindowsContext } from "../../../contexts/Windows";
 
 import PinnedContainer from "./PinnedContainer";
 import SortableItem from "./SortableItem";
